Изменение структуры парсеров
