+++ link-parser/src/main/java/ru/tinkoff/edu/java/linkparser/github/GitHubChainLinkParser.java
@@ -1,6 +1,6 @@
 package ru.tinkoff.edu.java.linkparser.github;
 
-import ru.tinkoff.edu.java.linkparser.common.AbstractLinkParser;
+import ru.tinkoff.edu.java.linkparser.common.AbstractChainLinkParser;
 import ru.tinkoff.edu.java.linkparser.github.payload.GitHubPayload;
 
 import java.util.Objects;
@@ -9,14 +9,14 @@ import java.util.regex.Pattern;
 /**
  * Парсер ссылок ресурса github.com
  */
-public final class GitHubLinkParser<T> extends AbstractLinkParser {
+public final class GitHubChainLinkParser<T> extends AbstractChainLinkParser {
 
     public static final Pattern PATTERN_GITHUB_COM_USER_REPO =
             Pattern.compile("^[a-z]+://([^.]+(?<!github)[.])?github[.]com/(?<user>[^/]+)/(?<repo>[^/]+)");
 
     private final Result.Processor<T, GitHubPayload> resultProcessor;
 
-    public GitHubLinkParser(final Result.Processor<T, GitHubPayload> resultProcessor) {
+    public GitHubChainLinkParser(final Result.Processor<T, GitHubPayload> resultProcessor) {
         this.resultProcessor = Objects.requireNonNull(resultProcessor);
     }
