+++ link-parser/src/main/java/ru/tinkoff/edu/java/linkparser/stackoverflow/StackOverflowChainLinkParser.java
@@ -1,6 +1,6 @@
 package ru.tinkoff.edu.java.linkparser.stackoverflow;
 
-import ru.tinkoff.edu.java.linkparser.common.AbstractLinkParser;
+import ru.tinkoff.edu.java.linkparser.common.AbstractChainLinkParser;
 import ru.tinkoff.edu.java.linkparser.stackoverflow.payload.StackOverflowPayload;
 
 import java.util.Objects;
@@ -9,14 +9,14 @@ import java.util.regex.Pattern;
 /**
  * Парсер ссылок ресурса stackoverflow.com
  */
-public final class StackOverflowLinkParser<T> extends AbstractLinkParser {
+public final class StackOverflowChainLinkParser<T> extends AbstractChainLinkParser {
 
     public static final Pattern PATTERN_STACKOVERFLOW_QUESTION_ID =
             Pattern.compile("^[a-z]+://stackoverflow[.]com/questions/(?<questionId>\\d+)");
 
     private final Result.Processor<T, StackOverflowPayload> resultProcessor;
 
-    public StackOverflowLinkParser(final Result.Processor<T, StackOverflowPayload> resultProcessor) {
+    public StackOverflowChainLinkParser(final Result.Processor<T, StackOverflowPayload> resultProcessor) {
         this.resultProcessor = Objects.requireNonNull(resultProcessor);
     }
